se agrega validacion de noticia al eliminar una categoria

+++ src/controllers/categoria.controllers.js
@@ -1,4 +1,5 @@
 import Categoria from "../models/categoria";
+import Noticia from "../models/noticia";
 const categoriaCtrl = {};
 
 categoriaCtrl.nuevaCategoria = async (req, res) => {
@@ -46,6 +47,13 @@ categoriaCtrl.listarCategorias = async (req, res) => {
 
 categoriaCtrl.eliminarCategoria = async (req, res) => {
   try {
+    const categoria = await Categoria.findById(req.params.id);
+    const arregloNoticias = await Noticia.find({categoria: categoria.nombre}).collation({locale: "es", strength: 2});
+    if (arregloNoticias.length > 0) {
+      return res.status(500).json({
+        mensaje: "Debe eliminar primero las noticias de esta categoria",
+      });
+    }
     await Categoria.findByIdAndDelete(req.params.id);
     res.status(200).json({
       mensaje: "Se elimnino la categoria correctamente",
